+++ docs/release_notes.md
@@ -4,6 +4,7 @@
 
 ## unreleased
 * __参数`seed`的类型由必须为`int`改为可为任一`hashable`__
+* __提供参数`color`对所有[Pillow Color Name](https://pillow.readthedocs.io/en/5.2.x/reference/ImageColor.html#color-names)的支持__
 
 
 ## v1.3.0 (2018-7-23)